refactor(server): split architecture boundaries
